split TestTestUnicodeInvariants into multiple classes
